🚧 investigate missing species & links
I realized the number of preys and predators did not match the number of mammals. Turns out it's because Hippopotamus amphibius and Loxodonta africana (hippos and elephants) have no predators in our network and only have plants for preys, for which we have no data. Therefore, hippos and elephants have no interactions and are not included in our range updating analyses